server/pyproject.toml
[project]
name = "service-desk-api"
version = "0.1.0"
description = "FastAPI backend for the Service Desk platform."
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlmodel",
  "passlib[bcrypt]",
  "python-dotenv",
  "python-jose[cryptography]",
  "pydantic-settings",
  "python-multipart>=0.0.20",
  "crewai",
  "crewai-tools>=0.49.0",
]


[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["api*", "core*", "db*"]
